Move loaders.py into individual files under loaders/
